+++ examples/automation.py
@@ -5,26 +5,25 @@ import asyncio
 import json
 import time
 
-from examples.common import FAUCET_URL, NODE_URL
+from examples.common import RPC_NODE_URL
 from supra_sdk.account import Account
 from supra_sdk.account_address import AccountAddress
-from supra_sdk.async_client import ClientConfig, FaucetClient, RestClient
 from supra_sdk.bcs import Serializer
+from supra_sdk.clients.rest import SupraClient, SupraClientConfig
 from supra_sdk.transactions import EntryFunction, TransactionArgument
 
 
 async def main():
-    client_config = ClientConfig(
+    supra_client_config = SupraClientConfig(
         expiration_ttl=300,
         gas_unit_price=100,
         max_gas_amount=100_000,
     )
-    rest_client = RestClient(NODE_URL, client_config)
-    faucet_client = FaucetClient(FAUCET_URL, rest_client)
+    supra_client = SupraClient(RPC_NODE_URL, supra_client_config)
     sender = Account.generate()
     print(f"Automation task creator account address: {sender.address()}")
 
-    await faucet_client.faucet(sender.address())
+    await supra_client.faucet(sender.address())
     print(f"Faucet on {sender.address()} done")
 
     print("Registering automation tasK")
@@ -37,17 +36,17 @@ async def main():
             TransactionArgument(1, Serializer.u64),
         ],
     )
-    task_registration_tx_hash = await rest_client.register_automation_task(
+    task_registration_tx_hash = await supra_client.register_automation_task(
         sender,
         automated_function,
         9,
         100,
-        4 * (10**8),
+        1 * (10**8),
         int(time.time()) + 7200,
         [],
     )
 
-    task_registration_tx_data = await rest_client.transaction_by_hash(
+    task_registration_tx_data = await supra_client.transaction_by_hash(
         task_registration_tx_hash
     )
     task_index = None
@@ -61,11 +60,10 @@ async def main():
     print(f"Task registered successfully, task_id: {task_index}")
 
     print(f"Stopping task, task_index: {task_index}")
-    await rest_client.stop_automation_tasks(sender, [int(task_index)])
+    await supra_client.stop_automation_tasks(sender, [int(task_index)])
     print(f"Task stopped successfully, task_index: {task_index}")
 
-    await rest_client.close()
-    await faucet_client.close()
+    await supra_client.close()
 
 
 if __name__ == "__main__":
