+++ examples/hello_blockchain.py
@@ -14,20 +14,20 @@ import asyncio
 import sys
 from typing import Any
 
+from examples.common import RPC_NODE_URL
 from supra_sdk.account import Account
 from supra_sdk.account_address import AccountAddress
-from supra_sdk.async_client import ApiError, FaucetClient, RestClient
 from supra_sdk.bcs import Serializer
+from supra_sdk.clients.api_client import ApiError
+from supra_sdk.clients.rest import SupraClient
 from supra_sdk.transactions import (
     EntryFunction,
     TransactionArgument,
     TransactionPayload,
 )
 
-from .common import FAUCET_URL, NODE_URL
 
-
-class HelloBlockchainClient(RestClient):
+class HelloBlockchainClient(SupraClient):
     async def get_message(
         self, contract_address: AccountAddress, account_address: AccountAddress
     ) -> dict[str, Any] | None:
@@ -65,34 +65,31 @@ async def main(contract_address: AccountAddress):
     print(f"Alice account address: {alice.address()}")
     print(f"Bob account address: {bob.address()}")
 
-    rest_client = HelloBlockchainClient(NODE_URL)
-    faucet_client = FaucetClient(FAUCET_URL, rest_client)
-
-    await faucet_client.faucet(alice.address())
-    await faucet_client.faucet(bob.address())
+    supra_client = HelloBlockchainClient(RPC_NODE_URL)
+    await supra_client.faucet(alice.address())
+    await supra_client.faucet(bob.address())
 
     print("\n=== Testing Alice ===")
-    message = await rest_client.get_message(contract_address, alice.address())
+    message = await supra_client.get_message(contract_address, alice.address())
     print(f"Initial value: {message}")
     print('Setting the message to "Hello, Blockchain"')
-    await rest_client.set_message(contract_address, alice, "Hello, Blockchain")
-    message = await rest_client.get_message(contract_address, alice.address())
+    await supra_client.set_message(contract_address, alice, "Hello, Blockchain")
+    message = await supra_client.get_message(contract_address, alice.address())
     print(f"New value: {message}")
 
     print("\n=== Testing Bob ===")
-    message = await rest_client.get_message(contract_address, bob.address())
+    message = await supra_client.get_message(contract_address, bob.address())
     print(f"Initial value: {message}")
     print('Setting the message to "Hello, Blockchain"')
-    await rest_client.set_message(contract_address, bob, "Hello, Blockchain")
-    message = await rest_client.get_message(contract_address, bob.address())
+    await supra_client.set_message(contract_address, bob, "Hello, Blockchain")
+    message = await supra_client.get_message(contract_address, bob.address())
     print(f"New value: {message}")
 
-    await rest_client.close()
-    await faucet_client.close()
+    await supra_client.close()
 
 
 if __name__ == "__main__":
     assert len(sys.argv) == 2, "Expecting the contract address"
-    contract_address = sys.argv[1]
+    contract_address_str = sys.argv[1]
 
-    asyncio.run(main(AccountAddress.from_str(contract_address)))
+    asyncio.run(main(AccountAddress.from_str(contract_address_str)))
