+++ supra_sdk/clients/rest/__init__.py
@@ -1,0 +1,6 @@
+# Copyright © Supra
+# SPDX-License-Identifier: Apache-2.0
+
+from supra_sdk.clients.rest.supra_client import SupraClient, SupraClientConfig
+
+__all__ = ["SupraClient", "SupraClientConfig"]
