+++ supra_sdk/supra_tokenv1_client.py
@@ -6,7 +6,8 @@ from typing import Any
 
 from supra_sdk.account import Account
 from supra_sdk.account_address import AccountAddress
-from supra_sdk.async_client import ApiError, RestClient
+from supra_sdk.clients.rest import SupraClient
+from supra_sdk.clients.api_client import ApiError
 from supra_sdk.bcs import Serializer
 from supra_sdk.transactions import (
     EntryFunction,
@@ -20,9 +21,9 @@ U64_MAX = 18446744073709551615
 class SupraTokenV1Client:
     """A wrapper around reading and mutating SupraTokens also known as Token Objects"""
 
-    client: RestClient
+    client: SupraClient
 
-    def __init__(self, client: RestClient):
+    def __init__(self, client: SupraClient):
         self.client = client
 
     async def create_collection(
